Accept 3 or more command line arguments.
As long as there are at least 3 arguments from the command line (with the 1st argument being the name of the script), the script will use the last 2 arguments as the reference repository and the repository shadow respectively, while all the preceding arguments are fed to the 'fossil init' command that is run. So arguments that 'fossil init' accepts such as -A can be passed through the script.